+++ provider/slack_channel_data_test.go
@@ -29,7 +29,8 @@ func TestOfflineSlackChannelsReadMemberID(t *testing.T) {
 		return
 	}
 	r := schema.TestResourceDataRaw(t, dataSourceSlackChannel().Schema, map[string]interface{}{
-		"slack_channel_id": "C01010101Z",
+		"slack_channel_id":   "C01010101Z",
+		"slack_channel_name": "team-rocket",
 	})
 
 	d := dataFireHydrantSlackChannelRead(context.Background(), r, c)